feat: add top and regular sponsors
