+++ src/views/Home/components/Sponsors/StartupSponsor.tsx
@@ -1,0 +1,94 @@
+import data from "../../../../data/2023.json";
+import {
+  StyledFlexGrow,
+  StyledLogos,
+  StyledSeparator,
+  StyledSlashes,
+  StyledSponsorIconNano,
+  StyledSponsorItemContainer,
+  StyledSponsorLogosContainer,
+  StyledSponsorTitleContainer,
+  StyledSponsorTitleMargin,
+  StyledSponsorTitleSlashesContainer,
+} from "./Sponsors.style";
+import SponsorBadge from "./SponsorBadge";
+import { Color } from "../../../../styles/colors";
+import { BIG_BREAKPOINT } from "../../../../constants/BreakPoints";
+import { buildSlashes } from "./Sponsors";
+import { useWindowSize } from "react-use";
+import { useCallback, useEffect, useState } from "react";
+
+export const StartupSponsor = () => {
+  const { width } = useWindowSize();
+  const [slashes, setSlashes] = useState("");
+  const [isHovered, setIsHovered] = useState<boolean>(false);
+
+  useEffect(() => {
+    const newSlashes = buildSlashes(2);
+
+    setSlashes(newSlashes);
+  }, [width]);
+
+  const handleHoverSponsorStartup = useCallback(() => setIsHovered(true), []);
+  const handleUnHoverSponsorStartup = useCallback(
+    () => setIsHovered(false),
+    []
+  );
+  let startupSponsors = data.sponsors.startup || [];
+
+  return (
+    <>
+      {data.sponsors.startup.length > 0 && (
+        <StyledSponsorItemContainer
+          className="SponsorItem startup"
+          onMouseEnter={handleHoverSponsorStartup}
+          onMouseLeave={handleUnHoverSponsorStartup}
+        >
+          <SponsorBadge
+            color={Color.DARK_BLUE}
+            position="right"
+            isVisible={isHovered}
+          />
+          <StyledSponsorTitleContainer className="SponsorTitle">
+            <StyledSponsorTitleSlashesContainer
+              color={Color.DARK_BLUE}
+              id="Slashes"
+            >
+              <StyledSlashes>{slashes}</StyledSlashes>
+              <StyledSeparator />
+
+              {width < BIG_BREAKPOINT && "STARTUP"}
+            </StyledSponsorTitleSlashesContainer>
+            {width >= BIG_BREAKPOINT && (
+              <StyledSponsorTitleSlashesContainer
+                color={
+                  isHovered && width >= BIG_BREAKPOINT
+                    ? Color.WHITE
+                    : Color.DARK_BLUE
+                }
+              >
+                <StyledSlashes>{slashes}</StyledSlashes>
+                STARTUP
+              </StyledSponsorTitleSlashesContainer>
+            )}
+            <StyledSponsorTitleMargin />
+          </StyledSponsorTitleContainer>
+
+          <StyledSponsorLogosContainer className="SponsorLogos">
+            <StyledFlexGrow />
+            <StyledLogos position="right">
+              {startupSponsors.map((sponsor) => (
+                <a href={sponsor.website} target="_blank" rel={"noreferrer"}>
+                  <StyledSponsorIconNano
+                    key={sponsor.name}
+                    src={sponsor.image}
+                  />
+                </a>
+              ))}
+            </StyledLogos>
+          </StyledSponsorLogosContainer>
+        </StyledSponsorItemContainer>
+      )}
+    </>
+  );
+};
